% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{ph_authenticate_admin_username_password}
\alias{ph_authenticate_admin_username_password}
\title{authenticate}
\usage{
ph_authenticate_admin_username_password(app, api, username, password)
}
\arguments{
\item{app}{the app name}

\item{api}{the api to hit}

\item{username}{the username to use}

\item{password}{the password to use}
}
\value{
a json
}
\description{
authenticate
}
\examples{
\dontrun{
ph_authenticate_admin_username_password <- function(
   "eurocontrol-data",
   "/api/admins/auth-with-password",
   username,
   password)
}
}
