% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_records.R
\name{ph_list_records}
\alias{ph_list_records}
\title{list records}
\usage{
ph_list_records(app, api, collection, ..., perPage = 30)
}
\arguments{
\item{app}{the application name}

\item{api}{the api to hit}

\item{collection}{the collection}

\item{...}{other query params}

\item{perPage}{number of records per page}
}
\value{
a tibble
}
\description{
list records
}
\examples{
\dontrun{
ph_list_records(
  app,
  "/api/collections/",
  collection = "nw_traffic",
  perPage = 1,
  sort = "-FLIGHT_DATE",
  skipTotal = 1)

# get only subset of columns
ph_list_records(
  app,
  "/api/collections/",
  collection = "nw_traffic",
  sort = "-FLIGHT_DATE",
  fileds = "id,FLIGHT_DATE,DAY_TFC"
  )
}
}
