% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_record.R
\name{ph_update_record}
\alias{ph_update_record}
\title{Update record}
\usage{
ph_update_record(app, api, collection, id, token = NULL, ...)
}
\arguments{
\item{app}{the app name}

\item{api}{the api}

\item{collection}{the collection name or id}

\item{id}{the record id}

\item{token}{the authentication token, depending on collection's \code{updateRule}}

\item{...}{the query params as supported by
\href{https://pocketbase.io/docs/api-records/#update-record}{PocketBase View API}}
}
\value{
a response
}
\description{
Update record
}
\examples{
\dontrun{
delete_record(app, api, id, token)
}
}
