% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pockethostr-package.R
\docType{package}
\name{pockethostr-package}
\alias{pockethostr}
\alias{pockethostr-package}
\title{pockethostr: Pockethost for R}
\description{
This package provides an R API for pockethost app instances.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eurocontrol/pockethostr}
  \item \url{https://euctrl-pru.github.io/pockethostr/}
}

}
\author{
\strong{Maintainer}: Enrico Spinielli \email{enrico.spinielli@gmail.com} (\href{https://orcid.org/0000-0001-8584-9131}{ORCID})

}
\keyword{internal}
