% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_record.R
\name{ph_delete_record}
\alias{ph_delete_record}
\title{Delete record}
\usage{
ph_delete_record(app, api, collection, id, token = NULL)
}
\arguments{
\item{app}{the app name}

\item{api}{the api}

\item{collection}{the collection name or id}

\item{id}{the record id}

\item{token}{the authentication token, depending on collection's \code{deleteRule}}
}
\value{
a response
}
\description{
Delete record
}
\examples{
\dontrun{
delete_record(app, api, id, token)
}
}
