% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_record.R
\name{ph_update_record}
\alias{ph_update_record}
\title{Update record}
\usage{
ph_update_record(app, api, collection, id, body, token = NULL, ...)
}
\arguments{
\item{app}{the app name}

\item{api}{the api}

\item{collection}{the collection name or id}

\item{id}{the record id}

\item{body}{the body}

\item{token}{the authentication token, depending on collection's \code{updateRule}}

\item{...}{the query params as supported by
\href{https://pocketbase.io/docs/api-records/#update-record}{PocketBase View API}}
}
\value{
a response
}
\description{
Update record
}
\examples{
\dontrun{
adm <- ph_authenticate_admin_username_password(
  "eurocontrol-data-test",
  "/api/admins/auth-with-password",
  username,
  password)
  bd <- list(
    DAY_TFC= 121,
    FLIGHT_DATE= "2024-05-01 00:00:00.000Z"
    )
 id <- "lufkj3346ygx9a2"
 ph_update_record(
  app = "eurocontrol-data-test",
  api = "/api/collections",
  collection = "nw_traffic",
  token = adm$token,
  body = bd)
}
}
