% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_record.R
\name{ph_view_record}
\alias{ph_view_record}
\title{View record}
\usage{
ph_view_record(app, api, collection, id, ...)
}
\arguments{
\item{app}{the app name}

\item{api}{the api}

\item{collection}{the collection}

\item{id}{the record id}

\item{...}{other query params}
}
\value{
a response
}
\description{
View record
}
\examples{
\dontrun{
ph_view_record(app, api, id, fileds = "COL1,COL3")
}
}
